requirements.txt
pyserial>=3.5
psutil>=5.8.0
